% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_aucs.R
\name{calc_auprc}
\alias{calc_auprc}
\title{Calculate AUPRC}
\usage{
calc_auprc(pred, bin_outcomes)
}
\arguments{
\item{pred}{Predictions of the trained model on the test data}

\item{bin_outcomes}{Binary outcome vector}
}
\value{
Area under the precision-recall curve
}
\description{
Calculate AUPRC
}
\author{
Zena Lapp, \email{zenalapp@umich.edu}
}
