% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_aucs.R
\name{calc_auroc}
\alias{calc_auroc}
\title{Title}
\usage{
calc_auroc(pred, bin_outcome)
}
\arguments{
\item{pred}{TODO}

\item{bin_outcome}{TODO}
}
\value{

}
\description{
Title
}
\author{
Zena Lapp, \email{zenalapp@umich.edu}
}
