% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{calc_baseline_precision}
\alias{calc_baseline_precision}
\title{Calculate the fraction of positives, i.e. baseline precision}
\usage{
calc_baseline_precision(dataset, outcome_colname, pos_outcome)
}
\arguments{
\item{dataset}{Dataframe with an outcome variable and other columns as features.}

\item{outcome_colname}{Column name as a string of the outcome variable
(default \code{NULL}; the first column will be chosen automatically).}

\item{pos_outcome}{the positive outcome from \code{outcome_colname},
e.g. "cancer" for the \code{otu_mini_bin} dataset.}
}
\value{
the baseline precision based on the fraction of positives
}
\description{
Calculate the fraction of positives, i.e. baseline precision
}
\examples{

calc_baseline_precision(otu_mini_bin, 'dx', 'cancer')

data.frame(y = c('a','b','a','b')) \%>\% calc_baseline_precision('y', 'a')

}
\author{
Kelly Sovacool, \email{sovacool@umich.edu}
}
