% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{calc_mean_perf}
\alias{calc_mean_perf}
\title{Generic function to calculate mean performance curves for multiple models}
\usage{
calc_mean_perf(sensspec_dat, group_var = specificity, sum_var = sensitivity)
}
\arguments{
\item{sensspec_dat}{data frame created by concatenating results of
\code{calc_model_sensspec()} for multiple models.}

\item{group_var}{variable to group by (e.g. specificity or recall).}

\item{sum_var}{variable to summarize (e.g. sensitivity or precision).}
}
\value{
data frame with mean & standard deviation of \code{sum_var} summarized over \code{group_var}
}
\description{
Generic function to calculate mean performance curves for multiple models
}
\author{
Courtney Armour

Kelly Sovacool
}
\keyword{internal}
