% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_importance.R
\name{calc_perm_perf}
\alias{calc_perm_perf}
\title{Calculate performance for one permutation of one feature}
\usage{
calc_perm_perf(
  feat_group,
  perm,
  test_data,
  trained_model,
  outcome_colname,
  perf_metric_function,
  perf_metric_name,
  class_probs,
  progbar = NULL
)
}
\arguments{
\item{feat_group}{group of correlated features as a character}

\item{perm}{permutation number}

\item{test_data}{Held out test data: dataframe of outcome and features.}

\item{trained_model}{Trained model from \code{\link[caret:train]{caret::train()}}.}

\item{outcome_colname}{Column name as a string of the outcome variable
(default \code{NULL}; the first column will be chosen automatically).}

\item{perf_metric_function}{Function to calculate the performance metric to
be used for cross-validation and test performance. Some functions are
provided by caret (see \code{\link[caret:postResample]{caret::defaultSummary()}}).
Defaults: binary classification = \code{twoClassSummary},
multi-class classification = \code{multiClassSummary},
regression = \code{defaultSummary}.}

\item{perf_metric_name}{The column name from the output of the function
provided to perf_metric_function that is to be used as the performance metric.
Defaults: binary classification = \code{"ROC"},
multi-class classification = \code{"logLoss"},
regression = \code{"RMSE"}.}

\item{class_probs}{Whether to use class probabilities (TRUE for categorical outcomes, FALSE for numeric outcomes).}
}
\value{

}
\description{
Calculate performance for one permutation of one feature
}
\examples{

}
\author{
Kelly Sovacool, \email{sovacool@umich.edu}
}
