% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_pvalue}
\alias{calc_pvalue}
\title{Calculate the p-value for a permutation test}
\usage{
calc_pvalue(vctr, test_stat)
}
\arguments{
\item{vctr}{vector of statistics}

\item{test_stat}{the test statistic}
}
\value{
the number of observations in \code{vctr} that are greater than
\code{test_stat} divided by the number of observations in \code{vctr}
}
\description{
compute Monte Carlo p-value with correction
based on formula from Page 158 of 'Bootstrap methods and their application'
By Davison & Hinkley 1997
}
\author{
Kelly Sovacool \email{sovacool@umich.edu}
}
\keyword{internal}
