% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{change_to_num}
\alias{change_to_num}
\title{Change columns to numeric if possible}
\usage{
change_to_num(features)
}
\arguments{
\item{features}{dataframe of features for machine learning}
}
\value{
dataframe with numeric columns where possible
}
\description{
Change columns to numeric if possible
}
\examples{
class(change_to_num(data.frame(val = c("1", "2", "3")))[[1]])
}
\author{
Zena Lapp, \email{zenalapp@umich.edu}
}
