% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_cat_feats}
\alias{check_cat_feats}
\title{Check if any features are categorical}
\usage{
check_cat_feats(feats)
}
\arguments{
\item{feats}{features}
}
\description{
Check if any features are categorical
}
\examples{
\dontrun{
check_cat_feats(otu_mini_bin)
}
}
\author{
Zena Lapp, \email{zenalapp@umich.edu}
}
\keyword{internal}
