% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_corr_thresh}
\alias{check_corr_thresh}
\title{check that corr_thresh is either NULL or a number between 0 and 1}
\usage{
check_corr_thresh(corr_thresh)
}
\arguments{
\item{corr_thresh}{correlation threshold}
}
\description{
check that corr_thresh is either NULL or a number between 0 and 1
}
\examples{
check_corr_thresh(1)
check_corr_thresh(0.8)
check_corr_thresh(2019)
check_corr_thresh(NULL)
}
\author{
Zena Lapp, \email{zenalapp@umich.edu}
}
\keyword{internal}
