% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_features}
\alias{check_features}
\title{Title}
\usage{
check_features(features, check_missing = TRUE)
}
\arguments{
\item{features}{features for machine learning}

\item{check_missing}{check whether the features have missing data (default: TRUE)}
}
\value{

}
\description{
Title
}
