% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_group}
\alias{check_group}
\title{Check group}
\usage{
check_group(dataset, group, kfold)
}
\arguments{
\item{dataset}{dataframe with an outcome variable and other columns as features}

\item{group}{vector of groups to keep together when splitting the data into train and test sets, and for cross-validation; length matches the number of rows in the dataset (default: no groups)}

\item{kfold}{fold number for k-fold cross-validation (default: 5)}
}
\value{

}
\description{
Check group
}
\examples{
check_group(mikRopML::otu_mini, sample(LETTERS,nrow(mikRopML::otu_mini),replace=TRUE),kfold=2)
}
