% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning_grid.R
\name{check_hyperparams_df}
\alias{check_hyperparams_df}
\title{Ensure the hyperparameters dataframe has a valid format}
\usage{
check_hyperparams_df(hyperparams_df, method_name)
}
\arguments{
\item{hyperparams_df}{dataframe of hyperparameters (default: default_hyperparams)}

\item{method_name}{method name (regLogistic, svmRadial, rpart2, rf, xgbTree)}
}
\value{
hyperparams df, filtered by method if needed
}
\description{
Ensure the hyperparameters dataframe has a valid format
}
\examples{
check_hyperparams_df(default_hyperparams, "regLogistic")
}
\author{
Kelly Sovacool, \email{sovacool@umich.edu}
}
