% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_kfold}
\alias{check_kfold}
\title{Check that kfold is an integer of reasonable size}
\usage{
check_kfold(kfold, dataset)
}
\arguments{
\item{kfold}{Fold number for k-fold cross-validation (default: \code{5}).}

\item{dataset}{Data frame with an outcome variable and other columns as features.}
}
\description{
Check that kfold is an integer of reasonable size
}
\examples{
\dontrun{
check_kfold(5, otu_small)
}
}
\author{
Kelly Sovacool, \email{sovacool@umich.edu}
}
\keyword{internal}
