% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_method}
\alias{check_method}
\title{Check if the method is supported. If not, throws error.}
\usage{
check_method(method, hyperparameters)
}
\arguments{
\item{method}{ML method. Options: \code{c("glmnet", "rf", "rpart2", "svmRadial", "xgbTree")}.
\itemize{
\item glmnet: linear, logistic, or multiclass regression
\item rf: random forest
\item rpart2: decision tree
\item svmRadial: support vector machine
\item xgbTree: xgboost
}}

\item{hyperparameters}{Dataframe of hyperparameters (default \code{NULL}; sensible
defaults will be chosen automatically).}
}
\description{
Check if the method is supported. If not, throws error.
}
\examples{
\dontrun{
check_method("rf")
}
}
\author{
Kelly Sovacool, \email{sovacool@umich.edu}
}
\keyword{internal}
