% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_ntree}
\alias{check_ntree}
\title{Check ntree}
\usage{
check_ntree(ntree)
}
\description{
Check ntree
}
\examples{
\dontrun{
check_ntree(NULL)
}
}
\author{
Zena Lapp, \email{zenalapp@umich.edu}
}
\keyword{internal}
