% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_outcome_value}
\alias{check_outcome_value}
\title{Check that the outcome variable is valid. Pick outcome value if necessary.}
\usage{
check_outcome_value(dataset, outcome_colname, pos_outcome = NULL)
}
\arguments{
\item{dataset}{Data frame with an outcome variable and other columns as features.}

\item{outcome_colname}{Column name as a string of the outcome variable
(default \code{NULL}; the first column will be chosen automatically).}
}
\value{
outcome value
}
\description{
Check that the outcome variable is valid. Pick outcome value if necessary.
}
\examples{
\dontrun{
check_outcome_value(otu_small, "dx", "cancer")
}
}
\author{
Zena Lapp, \email{zenalapp@umich.edu}

Kelly Sovacool, \email{sovacool@umich.edu}
}
\keyword{internal}
