% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_feature_importance.R
\name{check_package_installed}
\alias{check_package_installed}
\title{Check if package is installed}
\usage{
check_package_installed(package_name)
}
\arguments{
\item{package_name}{name of package to check}
}
\value{
boolean - whether package is installed (T) or not F).
}
\description{
Check if package is installed
}
\examples{
check_package_installed("base"); check_package_installed("asdf")
}
