% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_perf_metric_function}
\alias{check_perf_metric_function}
\title{Check perf_metric_function is NULL or a function}
\usage{
check_perf_metric_function(perf_metric_function)
}
\arguments{
\item{perf_metric_function}{performance metric function}
}
\description{
Check perf_metric_function is NULL or a function
}
\examples{
check_perf_metric_function(NULL)
}
\author{
Zena Lapp, \email{zenalapp@umich.edu}
}
\keyword{internal}
