% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_perf_metric_name}
\alias{check_perf_metric_name}
\title{Check perf_metric_name is NULL or a function}
\usage{
check_perf_metric_name(perf_metric_name)
}
\arguments{
\item{perf_metric_name}{performance metric function}
}
\description{
Check perf_metric_name is NULL or a function
}
\examples{
\dontrun{
check_perf_metric_name(NULL)
}
}
\author{
Zena Lapp, \email{zenalapp@umich.edu}
}
\keyword{internal}
