% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_permute}
\alias{check_permute}
\title{Check that permute is a logical}
\usage{
check_permute(permute)
}
\description{
Check that permute is a logical
}
\examples{
\dontrun{
do_permute <- TRUE
check_permute(do_permute)
}
}
\author{
Kelly Sovacool, \email{sovacool@umich.edu}
}
\keyword{internal}
