% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_remove_var}
\alias{check_remove_var}
\title{Check remove_var}
\usage{
check_remove_var(remove_var)
}
\arguments{
\item{remove_var}{Whether to remove variables with near-zero variance
(\code{'nzv'}; default), zero variance (\code{'zv'}), or none (\code{NULL}).}
}
\description{
Check remove_var
}
\examples{
check_remove_var(NULL)
}
\author{
Zena Lapp, \email{zenalapp@umich.edu}
}
\keyword{internal}
