% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_seed}
\alias{check_seed}
\title{check that the seed is either NA or a number}
\usage{
check_seed(seed)
}
\arguments{
\item{seed}{random seed}
}
\description{
check that the seed is either NA or a number
}
\examples{
check_seed(2019)
check_seed(NULL)
}
\author{
Kelly Sovacool, \email{sovacool@umich.edu}
}
\keyword{internal}
