% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_training_frac}
\alias{check_training_frac}
\title{Check that the training fraction is greater than 0 and less than or equal to 1}
\usage{
check_training_frac(frac)
}
\arguments{
\item{frac}{fraction (numeric)}
}
\description{
Check that the training fraction is greater than 0 and less than or equal to 1
}
\examples{
\dontrun{
check_training_frac(0.8)
}
}
\author{
Kelly Sovacool, \email{sovacool@umich.edu}
}
\keyword{internal}
