% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_training_indices}
\alias{check_training_indices}
\title{Check the validity of the training indices}
\usage{
check_training_indices(training_inds, dataset)
}
\arguments{
\item{training_inds}{vector of integers corresponding to samples for the training set}

\item{dataset}{data frame containing the entire dataset}
}
\description{
Check the validity of the training indices
}
\examples{
training_indices <- otu_small \%>\%
  nrow() \%>\%
  sample(., size = 160)
check_training_indices(training_indices, otu_small)
}
\author{
Kelly Sovacool, \email{sovacool@umich.edu}
}
\keyword{internal}
