% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{collapse_correlated_features}
\alias{collapse_correlated_features}
\title{Remove correlated features}
\usage{
collapse_correlated_features(features, group_neg_corr = TRUE)
}
\arguments{
\item{features}{features for ML}

\item{group_neg_corr}{whether to group negatively correlated features together (e.g. c(0,1) and c(1,0))}
}
\value{
features where perfectly correlated ones are collapsed
}
\description{
Remove correlated features
}
\examples{
<<<<<<< HEAD:man/collapse_correlated_features.Rd
collapse_correlated_features(mikRopML::otu_small[, 2:ncol(otu_small)])
=======
rm_corr_feats(mikropml::otu_small[, 2:ncol(otu_small)])
>>>>>>> master:man/rm_corr_feats.Rd
}
\author{
Zena Lapp, \email{zenalapp@umich.edu}
}
