% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{create_grouped_data_partition}
\alias{create_grouped_data_partition}
\title{Split into train and test set while splitting by groups}
\usage{
create_grouped_data_partition(groups, p)
}
\arguments{
\item{groups}{vector of groups. length must match the number of rows in the dataset.}

\item{p}{maximum fraction of data that goes to training (maybe less depending on groups sizes)}
}
\value{
vector of row indices for the training set
}
\description{
Split into train and test set while splitting by groups
}
\examples{
groups <- c("A", "B", "A", "B", "C", "C", "A", "A", "D")
set.seed(0)
train_ind <- create_grouped_data_partition(groups, 0.8)
}
\author{
Zena Lapp, {zenalapp@umich.edu}
}
