% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tuning_grid.R
\name{generate_tuning_grid}
\alias{generate_tuning_grid}
\title{Generate the Tuning Grid for Tuning Hyperparameters}
\usage{
generate_tuning_grid(model, hyperparameters)
}
\arguments{
\item{model}{Model name}

\item{hyperparameters}{Dataframe with hyperparameters. This dataframe should have the first column named as "param" which is the hyperparameter name and the second column "val" which are the values to be tested and third column "model" which is the model being used.}
}
\value{
List of the grid and the training package being used.
}
\description{
Generate the Tuning Grid for Tuning Hyperparameters
}
\examples{
generate_tuning_grid("L2_Logistic_Regression", default_hyperparams)
}
\author{
Begüm Topçuoğlu, \email{topcuoglu.begum@gmail.com}
}
