% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_aucs.R
\name{get_bin_outcome}
\alias{get_bin_outcome}
\title{Get binary outcome vector for calculating AUCs}
\usage{
get_bin_outcome(outcome_vec, outcome_samples)
}
\arguments{
\item{outcome_vec}{TODO}

\item{outcome_samples}{TODO}
}
\value{

}
\description{
Get binary outcome vector for calculating AUCs
}
\author{
Zena Lapp, \email{zenalapp@umich.edu}
}
