% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{get_caret_processed_df}
\alias{get_caret_processed_df}
\title{Get preprocessed dataframe for continuous variables}
\usage{
get_caret_processed_df(features, method)
}
\arguments{
\item{features}{dataframe of features for machine learning}

\item{method}{methods to preprocess the data, described in \code{caret::preProcess} (defaut: \code{c("center","scale")})}
}
\value{
processed matrix
}
\description{
Get preprocessed dataframe for continuous variables
}
\examples{
get_caret_processed_df(mikRopML::otu_small[, 2:ncol(otu_small)], c("center", "scale"))
}
