% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_hp_performance}
\alias{get_hp_performance}
\title{Get hyperparameter peformance metrics}
\usage{
get_hp_performance(trained_model)
}
\arguments{
\item{trained_model}{trained model (e.g. from \code{run_ml()})}
}
\value{
Named list:
\itemize{
\item \code{dat}: Dataframe of performance metric for each group of hyperparameters.
\item \code{params}: Hyperparameters tuned.
\item \code{metric}: Performance metric used.
}
}
\description{
Get hyperparameter peformance metrics
}
\examples{
get_hp_performance(otu_mini_results1$trained_model)
}
\author{
Zena Lapp, \email{zenalapp@umich.edu}

Kelly Sovacool \email{sovacool@umich.edu}
}
