% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning_grid.R
\name{get_hyperparams_list}
\alias{get_hyperparams_list}
\title{Split hyperparameters dataframe into named lists for each parameter}
\usage{
get_hyperparams_list(hyperparams_df)
}
\arguments{
\item{hyperparams_df}{dataframe of hyperparameters (default: default_hyperparams)}
}
\value{
named list of lists of hyperparameters
}
\description{
Split hyperparameters dataframe into named lists for each parameter
}
\examples{
get_hyperparams_list(default_hyperparams)
}
\author{
Begüm Topçuoğlu, \email{topcuoglu.begum@gmail.com}

Kelly Sovacool, \email{sovacool@umich.edu}
}
