% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tuning_grid.R
\name{get_method_hyperparams}
\alias{get_method_hyperparams}
\title{Filter hyperparameters by method and split into lists for each parameter}
\usage{
get_method_hyperparams(method, hyperparams_df)
}
\arguments{
\item{method}{method (regLogistic, svmRadial, rpart2, rf, xgbTree)}

\item{hyperparams_df}{dataframe with columns param, val, & model}
}
\value{
named list of lists of hyperparameters
}
\description{
Filter hyperparameters by method and split into lists for each parameter
}
\examples{
get_method_hyperparams("regLogistic", default_hyperparams)
}
\author{
Begüm Topçuoğlu, \email{topcuoglu.begum@gmail.com}

Kelly Sovacool, \email{sovacool@umich.edu}
}
