% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_metrics.R
\name{get_outcome_type}
\alias{get_outcome_type}
\title{Get outcome type}
\usage{
get_outcome_type(outcomes_vec)
}
\arguments{
\item{outcomes_vec}{Vector of outcomes.}
}
\value{
Outcome type (continuous, binary, or multiclass).
}
\description{
Get outcome type
}
\examples{
get_outcome_type(c(1, 2, 1))
get_outcome_type(c("a", "b", "b"))
get_outcome_type(c("a", "b", "c"))
}
\author{
Zena Lapp, \email{zenalapp@umich.edu}
}
