% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_outcome_value}
\alias{get_outcome_value}
\title{Get the outcome value of interest for AUC calculations}
\usage{
get_outcome_value(dataset, outcome_colname, method = "fewer")
}
\arguments{
\item{dataset}{Dataframe of input data}

\item{outcome_colname}{Column name of the outcome}

\item{method}{Method to choose outcome value of interest ("fewer", "first")}
}
\value{
outcome value of interest
}
\description{
Choose the outcome value of interest from the outcome column based on
which outcome has fewer rows or is the first row of the dataframe.
}
\examples{
get_outcome_value(otu_medium, "dx")
get_outcome_value(otu_medium, "dx", "first")
}
\author{
Kelly Sovacool, \email{sovacool@umich.edu}

Zena Lapp, \email{zenalapp@umich.edu}
}
