% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{get_partition_indices}
\alias{get_partition_indices}
\title{Select indices to partition the data into training & testing sets.}
\usage{
get_partition_indices(outcomes, training_frac = 0.8, groups = NULL)
}
\arguments{
\item{outcomes}{vector of outcomes}

\item{training_frac}{max fraction of data for the training set(default: 0.8)}

\item{groups}{vector of groups. length must match the number of rows in the dataset. (default: NULL)}
}
\value{
vector of row indices for the training set
}
\description{
If \code{groups} is \code{NULL}, uses \link[caret]{createDataPartition}.
Otherwises, uses \code{create_grouped_data_partition()}.
}
\details{
Set the seed prior to calling this function if you would like your data
partitions to be reproducible (recommended).
}
\author{
Kelly Sovacool, {sovacool@umich.edu}
}
