% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{get_perf_metric_fn}
\alias{get_perf_metric_fn}
\title{Get default performance metric function}
\usage{
get_perf_metric_fn(outcome_type)
}
\arguments{
\item{outcome_type}{Type of outcome (one of: \code{"continuous"},\code{"binary"},\code{"multiclass"}).}
}
\value{
Performance metric function.
}
\description{
Get default performance metric function
}
\examples{
get_perf_metric_fn("continuous")
get_perf_metric_fn("binary")
get_perf_metric_fn("multiclass")
}
\author{
Zena Lapp, \email{zenalapp@umich.edu}
}
