% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_metrics.R
\name{get_perf_metric_name}
\alias{get_perf_metric_name}
\title{Get default performance metric name}
\usage{
get_perf_metric_name(outcome_type)
}
\arguments{
\item{outcome_type}{type of outcome (one of: \code{"continuous"},\code{"binary"},\code{"multiclass"})}
}
\value{
performance metric name
}
\description{
Get default performance metric name
}
\examples{
get_perf_metric_name("continuous")
get_perf_metric_name("binary")
get_perf_metric_name("multiclass")
}
\author{
Zena Lapp, \email{zenalapp@umich.edu}
}
