% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_importance.R
\name{get_permuted_stats}
\alias{get_permuted_stats}
\title{Calculate summarized statistics from permutation tests}
\usage{
get_permuted_stats(perms_df, test_perf_value, method, perf_metric_name, seed)
}
\arguments{
\item{perms_df}{data frame of all permuted performance values, created with
furrr::future_map2_dfr() and calc_perm_perf()}

\item{method}{ML method.
Options: \code{c("glmnet", "rf", "rpart2", "svmRadial", "xgbTree")}.
\itemize{
\item glmnet: linear, logistic, or multiclass regression
\item rf: random forest
\item rpart2: decision tree
\item svmRadial: support vector machine
\item xgbTree: xgboost
}}

\item{perf_metric_name}{The column name from the output of the function
provided to perf_metric_function that is to be used as the performance metric.
Defaults: binary classification = \code{"ROC"},
multi-class classification = \code{"logLoss"},
regression = \code{"RMSE"}.}

\item{seed}{Random seed (default: \code{NA}).
Your results will only be reproducible if you set a seed.}
}
\value{
A tibble with performance metrics for when each feature (or group
of correlated features; \code{names}) is permuted (\code{perf_metric}), differences
between the actual test performance metric on and the permuted performance
metric (\code{perf_metric_diff}; test minus permuted performance), and the
p-value (\code{pvalue}: the probability of obtaining the actual performance
value under the null hypothesis). Features with a larger \code{perf_metric_diff}
are more important. The performance metric name (\code{perf_metric_name}) and
seed (\code{seed}) are also returned.
}
\description{
The \code{data.table} package is required.
}
\examples{

}
\author{
Kelly Sovacool, \email{sovacool@umich.edu}
}
