% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_aucs.R
\name{get_prediction}
\alias{get_prediction}
\title{Get predictions from trained model and test data}
\usage{
get_prediction(trained_model, test_data, outcome_value)
}
\arguments{
\item{trained_model}{trained model from caret}

\item{test_data}{dataframe of testing data}

\item{outcome_value}{Outcome value of interest}
}
\value{

}
\description{
Get predictions from trained model and test data
}
\author{
Zena Lapp, \email{zenalapp@umich.edu}
}
