% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tuning_grid.R
\name{get_tuning_grid}
\alias{get_tuning_grid}
\title{Generate the Tuning Grid for Tuning Hyperparameters}
\usage{
get_tuning_grid(method, hyperparams)
}
\arguments{
\item{method}{method name (regLogistic, svmRadial, rpart2, rf, xgbTree)}

\item{hyperparams}{Dataframe or named list of lists of hyperparameters}
}
\value{
The tuning grid
}
\description{
Generate the Tuning Grid for Tuning Hyperparameters
}
\examples{
get_tuning_grid("regLogistic", default_hyperparams)
}
\author{
Begüm Topçuoğlu, \email{topcuoglu.begum@gmail.com}

Kelly Sovacool, \email{sovacool@umich.edu}
}
