% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_by_group.R
\name{groupKMultiFolds}
\alias{groupKMultiFolds}
\title{Splitting into folds for cross-validation when using groups}
\usage{
groupKMultiFolds(group, kfold = 10, cv_times = 5)
}
\arguments{
\item{group}{equivalent to y in caret::createMultiFolds}

\item{kfold}{equivalent to k in caret::createMultiFolds}

\item{cv_times}{equivalent to cv_times in caret::createMultiFolds}
}
\value{
indices of folds for CV
}
\description{
Like \link[caret]{createMultiFolds} but still splitting by group using \link[caret]{groupKFold}. Code modified from \link[caret]{createMultiFolds}.
}
\examples{
set.seed(0)
group <- c('A','B','A','B','C','C','A','A','D')
folds <- groupKMultiFolds(group, kfold=2, cv_times=2)
}
