% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mikropml.R
\docType{package}
\name{mikropml}
\alias{mikropml}
\title{mikropml: User-Friendly R Package for Robust Machine Learning Pipelines}
\description{
\code{mikropml} implements robust machine learning pipelines using regression, support vector machines, decision trees, random forest, or gradient-boosted trees. The main functions are \code{preprocess_data()} to process your data prior to running machine learning, and \code{run_ml()} to run machine learning.
}
\section{Authors}{

\itemize{
\item Begüm D. Topçuoğlu (\href{0000-0003-3140-537X}{ORCID})
}
}

\section{See also}{

}

