% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_pipeline.R
\name{model_pipeline}
\alias{model_pipeline}
\title{Title}
\usage{
model_pipeline(
  dataset,
  model,
  split_number,
  outcome_colname = NA,
  outcome_value = NA,
  hyperparameters = NA,
  level = NA,
  permutation = TRUE
)
}
\arguments{
\item{dataset}{TODO}

\item{model}{TODO}

\item{split_number}{TODO}

\item{outcome_colname}{TODO}

\item{outcome_value}{TODO}

\item{hyperparameters}{TODO}

\item{level}{TODO}

\item{permutation}{TODO}
}
\value{

}
\description{
Title
}
\author{
Begüm Topçuoğlu, \email{topcuoglu.begum@gmail.com}
}
