% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mutate_all_types}
\alias{mutate_all_types}
\title{Mutate all columns with \code{utils::type.convert()}.`}
\usage{
mutate_all_types(dat)
}
\arguments{
\item{dat}{data.frame to convert}
}
\value{
data.frame with no factors
}
\description{
Turns factors into characters and numerics where possible.
}
\examples{
\dontrun{
dat <- data.frame(
  c1 = as.factor(c("a", "b", "c")),
  c2 = as.factor(1:3)
)
class(dat$c1)
class(dat$c2)
dat <- mutate_all_types(dat)
class(dat$c1)
class(dat$c2)
}
}
\author{
Kelly Sovacool, \email{sovacool@umich.edu}
}
\keyword{internal}
