% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{otu_data_preproc}
\alias{otu_data_preproc}
\title{Mini OTU abundance dataset - preprocessed}
\format{
An object of class \code{list} of length 3.
}
\usage{
otu_data_preproc
}
\description{
This is the result of running \code{preprocess_data("otu_mini_bin")}
}
\keyword{datasets}
