% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{otu_large}
\alias{otu_large}
\title{Large OTU abundance dataset}
\format{A data frame with 490 rows and 6,921 variables.
The \code{dx} column is the diagnosis: healthy or cancerous (colorectal).
All other columns are OTU relative abundances.}
\usage{
otu_large
}
\description{
A dataset containing relatives abundances of 6,920 OTUs for 490 human stool samples
}
\keyword{datasets}
