% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{otu_large_multi}
\alias{otu_large_multi}
\title{Large OTU abundance dataset with 3 classes}
\format{
A data frame with 490 rows and 6,921 variables.
The \code{dx} column is the colon cancer diagnosis: adenoma, carcinoma, or normal.
All other columns are OTU relative abundances.
}
\usage{
otu_large_multi
}
\description{
A dataset containing relatives abundances of 6,920 OTUs for 490 human stool samples
}
\keyword{datasets}
