% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{otu_med_results}
\alias{otu_med_results}
\title{Results from running the pipline with rpart2 on \code{otu_medium}}
\format{
An object of class \code{list} of length 4.
}
\usage{
otu_med_results
}
\description{
Results from running the pipline with rpart2 on \code{otu_medium}
}
\keyword{datasets}
