% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{otu_medium}
\alias{otu_medium}
\title{Medium-sized OTU abundance dataset}
\format{
A data frame with 60 rows and 6921 variables.
The \code{dx} column is the diagnosis: healthy or cancerous (colorectal).
All other columns are OTU relative abundances.
}
\usage{
otu_medium
}
\description{
A dataset containing relatives abundances of 6,920 OTUs for 60 human stool samples
}
\keyword{datasets}
