% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{otu_mini}
\alias{otu_mini}
\title{Mini OTU abundance dataset}
\format{
A data frame
The \code{dx} column is the diagnosis: healthy or cancerous (colorectal).
All other columns are OTU relative abundances.
}
\usage{
otu_mini
}
\description{
A dataset containing relatives abundances of OTUs for human stool samples.
This is a subset of the data provided in \code{extdata/otu_large.csv}.
}
\keyword{datasets}
