% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{otu_mini_bin}
\alias{otu_mini_bin}
\title{Mini OTU abundance dataset}
\format{
A data frame
The \code{dx} column is the diagnosis: healthy or cancerous (colorectal).
All other columns are OTU relative abundances.
}
\usage{
otu_mini_bin
}
\description{
A dataset containing relatives abundances of OTUs for human stool samples.
This is a subset of \code{otu_small}.
}
\keyword{datasets}
