% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{otu_mini_bin_results_glmnet}
\alias{otu_mini_bin_results_glmnet}
\title{Results from running the pipeline with L2 logistic regression on \code{otu_mini_bin} with feature importance and grouping}
\format{
An object of class \code{list} of length 4.
}
\usage{
otu_mini_bin_results_glmnet
}
\description{
Results from running the pipeline with L2 logistic regression on \code{otu_mini_bin} with feature importance and grouping
}
\keyword{datasets}
