% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{otu_mini_bin_results_svmRadial}
\alias{otu_mini_bin_results_svmRadial}
\title{Results from running the pipeline with svmRadial on \code{otu_mini_bin}}
\format{
An object of class \code{list} of length 4.
}
\usage{
otu_mini_bin_results_svmRadial
}
\description{
Results from running the pipeline with svmRadial on \code{otu_mini_bin}
}
\keyword{datasets}
