% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{otu_mini_cont_results_nocv}
\alias{otu_mini_cont_results_nocv}
\title{Results from running the pipeline with glmnet on \code{otu_mini_bin} with \code{Otu00001}
as the outcome column,
using a custom train control scheme that does not perform cross-validation}
\format{
An object of class \code{list} of length 4.
}
\usage{
otu_mini_cont_results_nocv
}
\description{
Results from running the pipeline with glmnet on \code{otu_mini_bin} with \code{Otu00001}
as the outcome column,
using a custom train control scheme that does not perform cross-validation
}
\keyword{datasets}
