% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{otu_mini_cv}
\alias{otu_mini_cv}
\title{cross validation on \code{train_data_mini} with grouped features}
\format{
An object of class \code{list} of length 27.
}
\usage{
otu_mini_cv
}
\description{
cross validation on \code{train_data_mini} with grouped features
}
\keyword{datasets}
