% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{otu_mini_cv2}
\alias{otu_mini_cv2}
\title{2-fold cross validation on \code{train_data_mini}}
\format{
An object of class \code{list} of length 27.
}
\usage{
otu_mini_cv2
}
\description{
2-fold cross validation on \code{train_data_mini}
}
\keyword{datasets}
