% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{otu_mini_cv2_grp}
\alias{otu_mini_cv2_grp}
\title{2-fold cross validation on \code{train_data_mini} with grouped features}
\format{
An object of class \code{list} of length 27.
}
\usage{
otu_mini_cv2_grp
}
\description{
2-fold cross validation on \code{train_data_mini} with grouped features
}
\keyword{datasets}
