% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{otu_mini_multi}
\alias{otu_mini_multi}
\title{Mini OTU abundance dataset with 3 categorical variables}
\format{
A data frame
The \code{dx} column is the colorectal cancerl diagnosis: adenoma, carcinoma, normal.
All other columns are OTU relative abundances.
}
\usage{
otu_mini_multi
}
\description{
A dataset containing relatives abundances of OTUs for human stool samples
}
\keyword{datasets}
