% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{otu_mini_multi_group}
\alias{otu_mini_multi_group}
\title{Groups for otu_mini_multi}
\format{
An object of class \code{character} of length 490.
}
\usage{
otu_mini_multi_group
}
\description{
Groups for otu_mini_multi
}
\keyword{datasets}
