% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{otu_mini_multi_results_glmnet}
\alias{otu_mini_multi_results_glmnet}
\title{Results from running the pipeline with glmnet on \code{otu_mini_multi} for multiclass outcomes}
\format{
An object of class \code{list} of length 4.
}
\usage{
otu_mini_multi_results_glmnet
}
\description{
Results from running the pipeline with glmnet on \code{otu_mini_multi} for multiclass outcomes
}
\keyword{datasets}
