% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{otu_mini_results1_grp}
\alias{otu_mini_results1_grp}
\title{Results from running the pipline with L2 logistic regression on \code{otu_mini} with feature grouping}
\format{
An object of class \code{list} of length 3.
}
\usage{
otu_mini_results1_grp
}
\description{
Results from running the pipline with L2 logistic regression on \code{otu_mini} with feature grouping
}
\keyword{datasets}
