% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{otu_mini_results1a}
\alias{otu_mini_results1a}
\title{Results from running the pipeline with L2 logistic regression with feature importance on \code{otu_mini}}
\format{
An object of class \code{list} of length 3.
}
\usage{
otu_mini_results1a
}
\description{
Results from running the pipeline with L2 logistic regression with feature importance on \code{otu_mini}
}
\keyword{datasets}
