% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{otu_mini_results5}
\alias{otu_mini_results5}
\title{Results from running the pipline with rpart2 on \code{otu_mini}}
\format{
An object of class \code{list} of length 3.
}
\usage{
otu_mini_results5
}
\description{
Results from running the pipline with rpart2 on \code{otu_mini}
}
\keyword{datasets}
