% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{otu_sm_cv5}
\alias{otu_sm_cv5}
\title{5-fold cross validation on \code{train_data_sm}}
\format{
An object of class \code{list} of length 27.
}
\usage{
otu_sm_cv5
}
\description{
5-fold cross validation on \code{train_data_sm}
}
\keyword{datasets}
