% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{otu_sm_results}
\alias{otu_sm_results}
\title{Results from running the pipline with L2 logistic regression on \code{otu_small}}
\format{
An object of class \code{list} of length 6.
}
\usage{
otu_sm_results
}
\description{
Results from running the pipline with L2 logistic regression on \code{otu_small}
}
\keyword{datasets}
