% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_importance.R
\name{permutation_importance}
\alias{permutation_importance}
\title{Title}
\usage{
permutation_importance(model, test_data, outcome, fewer_samples, level)
}
\arguments{
\item{model}{TODO}

\item{test_data}{TODO}

\item{outcome}{TODO}

\item{fewer_samples}{TODO}

\item{level}{TODO}
}
\value{

}
\description{
Title
}
\author{
Begüm Topçuoğlu, \email{topcuoglu.begum@gmail.com}
}
