% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_hp_performance}
\alias{plot_hp_performance}
\title{Plot hyperparameter performance metrics}
\usage{
plot_hp_performance(performance_df, param, metric)
}
\arguments{
\item{performance_df}{dataframe of hyperparameters and performance metric (e.g. from \code{get_hp_performance()} or \code{combine_hp_performance()})}

\item{param}{hyperparameter to be plotted}

\item{metric}{performance metric}
}
\description{
Plot hyperparameter performance metrics
}
\examples{
hp_metrics <- get_hp_performance(otu_mini_results1$trained_model)
plot_hp_performance(hp_metrics$dat,hp_metrics$params[1],hp_metrics$metric)
}
