% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{process_bin_feats}
\alias{process_bin_feats}
\title{Process binary features}
\usage{
process_bin_feats(features)
}
\arguments{
\item{features}{dataframe of features for machine learning}
}
\value{
list of two dataframes: binary (processed) and nonbinary features (unprocessed)
}
\description{
Process binary features
}
\examples{
process_bin_feats(mikRopML::otu_small[,2:ncol(otu_small)])
}
