% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{process_cat_feats}
\alias{process_cat_feats}
\title{Process categorical features}
\usage{
process_cat_feats(features, progbar = NULL)
}
\arguments{
\item{features}{dataframe of features for machine learning}

\item{progbar}{optional progress bar (default: \code{NULL})}
}
\value{
list of two dataframes: categorical (processed) and continuous features (unprocessed)
}
\description{
Process categorical features
}
\examples{
process_cat_feats(mikropml::otu_small[, 2:ncol(otu_small)])
}
\author{
Zena Lapp, \email{zenalapp@umich.edu}
}
\keyword{internal}
