% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{process_cont_feats}
\alias{process_cont_feats}
\title{Preprocess continuous features}
\usage{
process_cont_feats(features, method)
}
\arguments{
\item{features}{Dataframe of features for machine learning}

\item{method}{Methods to preprocess the data, described in
\code{\link[caret:preProcess]{caret::preProcess()}} (default: \code{c("center","scale")}, use \code{NULL} for
no normalization).}
}
\value{
dataframe of preprocessed features
}
\description{
Preprocess continuous features
}
\examples{
\dontrun{
process_cont_feats(mikropml::otu_small[, 2:ncol(otu_small)], c("center", "scale"))
}
}
\author{
Zena Lapp, \email{zenalapp@umich.edu}
}
\keyword{internal}
