% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{process_nonbin_feats}
\alias{process_nonbin_feats}
\title{Preprocess nonbinary features}
\usage{
process_nonbin_feats(features, method)
}
\arguments{
\item{features}{dataframe of features for machine learning}

\item{method}{methods to preprocess the data, described in \code{caret::preProcess} (defaut: \code{c("center","scale")})}
}
\value{
dataframe of preprocessed features
}
\description{
Preprocess nonbinary features
}
\examples{
process_nonbin_feats(mikRopML::otu_small[,2:ncol(otu_small)],c('center','scale','nzv'))
}
