% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{process_novar_feats}
\alias{process_novar_feats}
\title{Process features with no variation}
\usage{
process_novar_feats(features, progbar = NULL)
}
\arguments{
\item{features}{dataframe of features for machine learning}

\item{progbar}{optional progress bar (default: \code{NULL})}
}
\value{
list of two dataframes: features with variability (unprocessed) and without (processed)
}
\description{
Process features with no variation
}
\examples{
\dontrun{
process_novar_feats(mikropml::otu_small[, 2:ncol(otu_small)])
}
}
\author{
Zena Lapp, \email{zenalapp@umich.edu}
}
\keyword{internal}
