% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{radix_sort}
\alias{radix_sort}
\title{Call \code{sort()} with \code{method = 'radix'}}
\usage{
radix_sort(...)
}
\arguments{
\item{...}{All arguments forwarded to \code{sort()}.}
}
\value{
Whatever you passed in, now in a stable sorted order regardless of your locale.
}
\description{
THE BASE SORT FUNCTION USES A DIFFERENT METHOD DEPENDING ON YOUR LOCALE.
However, the order for the radix method is always stable.
}
\details{
see https://stackoverflow.com/questions/42272119/r-cmd-check-fails-devtoolstest-works-fine

\code{stringr::str_sort()} solves this problem with the \code{locale} parameter having
a default value, but I don't want to add that as another dependency.
}
\author{
Kelly Sovacool \email{sovacool@umich.edu}
}
\keyword{internal}
