% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_aucs.R
\name{recode_outcome}
\alias{recode_outcome}
\title{Recode the outcome column to a binary vector for calculating AUCs}
\usage{
recode_outcome(outcome_colname, outcome_value)
}
\arguments{
\item{outcome_colname}{Column name of the outcome variable}

\item{outcome_value}{Outcome value of interest}
}
\value{
Outcome column recoded as binary (1 = outcome of interest, 0 = other outcome)
}
\description{
Recode the outcome column to a binary vector for calculating AUCs
}
\author{
Zena Lapp, \email{zenalapp@umich.edu}

Kelly Sovacool, \email{sovacool@umich.edu}
}
