% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{contr.ltfr}
\alias{\%>\%}
\alias{.data}
\alias{!!}
\alias{:=}
\title{caret contr.ltfr}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{caret}{\code{\link[caret:contr.ltfr]{contr.ltfr()}}}

  \item{dplyr}{\code{\link[dplyr:\%>\%]{\%>\%}}}

  \item{rlang}{\code{\link[rlang::=]{:=()}}, \code{\link[rlang:!!]{!!()}}, \code{\link[rlang:.data]{.data}}}
}}

