% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_spaces}
\alias{replace_spaces}
\title{Replace spaces in all elements of a character vector with underscores}
\usage{
replace_spaces(x, new_char = "_")
}
\arguments{
\item{x}{a character vector}

\item{new_char}{the character to replace spaces (default: \verb{_})}
}
\value{
character vector with all spaces replaced with \code{new_char}
}
\description{
Replace spaces in all elements of a character vector with underscores
}
\examples{
dat <- data.frame(
  dx = c("outcome 1", "outcome 2", "outcome 1"),
  a = 1:3, b = c(5, 7, 1)
)
dat$dx <- replace_spaces(dat$dx)
dat
}
\author{
Kelly Sovacool, \email{sovacool@umich.edu}
}
