% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{rm_corr_feats}
\alias{rm_corr_feats}
\title{Remove correlated features}
\usage{
rm_corr_feats(features)
}
\arguments{
\item{features}{features for ML}
}
\value{
features where perfectly correlated ones are collapsed
}
\description{
Remove correlated features
}
\examples{
rm_corr_feats(mikRopML::otu_small[, 2:ncol(otu_small)])
}
\author{
Zena Lapp, \email{zenalapp@umich.edu}
}
