% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{rm_missing_outcome}
\alias{rm_missing_outcome}
\title{Remove missing outcome values}
\usage{
rm_missing_outcome(dataset, outcome_colname)
}
\arguments{
\item{dataset}{dataframe with an outcome variable and other columns as features}

\item{outcome_colname}{column name as a string of the outcome variable}
}
\value{
dataset with no missing outcomes
}
\description{
Remove missing outcome values
}
\examples{
rm_missing_outcome(mikRopML::otu_mini, "dx")

test_df <- mikRopML::otu_mini
test_df[1:100,'dx'] = NA
rm_missing_outcome(test_df, "dx")
}
