% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_model}
\alias{run_model}
\title{Title}
\usage{
run_model(
  seed,
  model,
  data_filename,
  hyperparameters,
  outcome,
  level,
  permutation
)
}
\arguments{
\item{seed}{TODO}

\item{model}{TODO}

\item{data_filename}{TODO}

\item{hyperparameters}{TODO}

\item{outcome}{TODO}

\item{level}{TODO}

\item{permutation}{TODO}
}
\value{

}
\description{
Title
}
\author{
Begüm Topçuoğlu, \email{topcuoglu.begum@gmail.com}
}
