% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pipeline.R
\name{run_pipeline}
\alias{run_pipeline}
\title{Run the machine learning pipeline}
\usage{
run_pipeline(
  dataset,
  method,
  outcome_colname = NA,
  outcome_value = NA,
  hyperparameters = mikRopML::default_hyperparams,
  permute = FALSE,
  nfolds = as.integer(5),
  training_frac = 0.8,
  seed = NA
)
}
\arguments{
\item{dataset}{dataframe with an outcome variable and other columns as features}

\item{method}{ML method ("regLogistic", "svmRadial", "rpart2", "rf", "xgbTree")}

\item{outcome_colname}{column name as a string of the outcome variable}

\item{outcome_value}{outcome value of interest as a string}

\item{hyperparameters}{dataframe of hyperparameters (default: default_hyperparams)}

\item{permute}{run permutation imporance (default: FALSE)}

\item{nfolds}{fold number for cross-validation (default: 5)}

\item{training_frac}{fraction size of data for training (default: 0.8)}

\item{seed}{random seed (default: NA)}
}
\value{
named list with results
}
\description{
Run the machine learning pipeline
}
\author{
Begüm Topçuoğlu, \email{topcuoglu.begum@gmail.com}

Zena Lapp, \email{zenalapp@umich.edu}

Kelly Sovacool, \email{sovacool@umich.edu}
}
