% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_apply}
\alias{select_apply}
\title{Use future apply if available}
\usage{
select_apply(fun = "apply")
}
\arguments{
\item{fun}{apply function to use (apply, lapply, sapply, etc.)}
}
\value{
output of apply function
}
\description{
Use future apply if available
}
\examples{
select_apply(fun = "sapply")
}
\author{
Zena Lapp, \email{zenalapp@umich.edu}
}
\keyword{internal}
