% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{sensspec}
\alias{sensspec}
\alias{get_model_sensspec}
\alias{calc_mean_roc}
\alias{calc_mean_prc}
\title{Calculate and summarize sensitivity, specificity, and precision.}
\usage{
get_model_sensspec(trained_model, test_data, outcome_colname, pos_outcome)

calc_mean_roc(sensspec_dat)

calc_mean_prc(sensspec_dat)
}
\arguments{
\item{trained_model}{Trained model from \code{\link[caret:train]{caret::train()}}.}

\item{test_data}{Held out test data: dataframe of outcome and features.}

\item{outcome_colname}{Column name as a string of the outcome variable
(default \code{NULL}; the first column will be chosen automatically).}

\item{pos_outcome}{the positive outcome from \code{outcome_colname},
e.g. "cancer" for the \code{otu_mini_bin} dataset.}

\item{sensspec_dat}{data frame by concatenating results of \code{get_model_sensspec()} for multiple models.}
}
\value{
data frame with summarized performance
}
\description{
These functions assume a binary outcome
}
\section{Functions}{
\itemize{
\item \code{get_model_sensspec()}: Get sensitivty, specificity, and precision for a model.

\item \code{calc_mean_roc()}: Calculate mean sensitivity over specificity

\item \code{calc_mean_prc()}: Calculate mean precision over recall

}}
\examples{
\dontrun{
# get cumulative performance for a single model
sensspec_1 <- get_model_sensspec(otu_mini_bin_results_glmnet$trained_model,
                   otu_mini_bin_results_glmnet$test_data,
                   'dx', 'cancer'
)
head(pred_1)

# get performance for multiple models
get_sensspec_seed <- function(seed) {
  ml_result <- run_ml(otu_mini_bin, 'glmnet', seed = seed)
  sensspec <- get_model_sensspec(ml_result$trained_model,
                                 ml_result$test_data,
                                 'dx', 'cancer') \%>\%
                                 mutate(seed = seed)
  return(sensspec)
}
sensspec_dat <- purrr::map_dfr(seq(100, 102), get_sensspec_seed)

# calculate mean sensitivity over specificity
roc_dat <- calc_mean_roc(sensspec_dat)
head(roc_dat)

# calculate mean precision over recall
prc_dat <- calc_mean_prc(sensspec_dat)
head(prc_dat)

# plot ROC & PRC
# TODO
}
}
\author{
Courtney Armour

Kelly Sovacool, \email{sovacool@umich.edu}
}
