% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparameters.R
\name{set_hparams_regLogistic}
\alias{set_hparams_regLogistic}
\title{Set hyperparameters for logistic regression models}
\usage{
set_hparams_regLogistic()
}
\value{
named list of hyperparameters
}
\description{
Set hyperparameters for logistic regression models
}
\examples{
set_hparams_regLogistic()
}
\author{
Kelly Sovacool, \email{sovacool@umich.edu}
}
