% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparameters.R
\name{set_hparams_rf}
\alias{set_hparams_rf}
\title{Set hyparameters for random forest models}
\usage{
set_hparams_rf(n_features)
}
\arguments{
\item{n_features}{number of features in the dataset}
}
\value{
named list of hyperparameters
}
\description{
Set hyparameters for random forest models
}
\examples{
\dontrun{
set_hparams_rf(16)
set_hparams_rf(2000)
set_hparams_rf(1)
}
}
\author{
Kelly Sovacool, \email{sovacool@umich.edu}
}
\keyword{internal}
