% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparameters.R
\name{set_hparams_svmRadial}
\alias{set_hparams_svmRadial}
\title{Set hyperparameters for SVM with radial kernel}
\usage{
set_hparams_svmRadial()
}
\value{
named list of hyperparameters
}
\description{
Set hyperparameters for SVM with radial kernel
}
\examples{
set_hparams_svmRadial()
}
\author{
Kelly Sovacool, \email{sovacool@umich.edu}
}
\keyword{internal}
