% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparameters.R
\name{set_hparams_xgbTree}
\alias{set_hparams_xgbTree}
\title{Set hyperparameters for SVM with radial kernel}
\usage{
set_hparams_xgbTree(n_samples)
}
\arguments{
\item{n_samples}{number of samples in the dataset}
}
\value{
named list of hyperparameters
}
\description{
Set hyperparameters for SVM with radial kernel
}
\examples{
set_hparams_xgbTree()
}
\author{
Kelly Sovacool, \email{sovacool@umich.edu}
}
\keyword{internal}
