% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{set_outcome_factor}
\alias{set_outcome_factor}
\title{Check or set outcome column to be a factor with \code{pos_class} as the first level}
\usage{
set_outcome_factor(dataset, outcome_colname, pos_class)
}
\arguments{
\item{dataset}{Data frame with an outcome variable and other columns as features.}

\item{outcome_colname}{Column name as a string of the outcome variable
(default \code{NULL}; the first column will be chosen automatically).}

\item{pos_class}{The positive class, i.e. which level of \code{outcome_colname} is
the event of interest. If the outcome is binary, either the
\code{outcome_colname} must be a factor with the first level being the positive
class, or \code{pos_class} must be set. (default: \code{NULL}).}
}
\value{
dataset, with the outcome column as a factor
}
\description{
Check or set outcome column to be a factor with \code{pos_class} as the first level
}
\examples{
dat <- data.frame('dx' = c('a','b','a','b','b','a'), feat = 1:6)
dat \%>\% set_outcome_factor('dx', 'a')
dat \%>\% set_outcome_factor('dx', 'b')
}
\author{
Kelly Sovacool, \email{sovacool@umich.edu}
}
\keyword{internal}
