% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{setup_parallel}
\alias{setup_parallel}
\title{Register a multiprocessing plan with \code{ncores}}
\usage{
setup_parallel(ncores)
}
\arguments{
\item{`ncores`}{number of cores for multiprocessing (default: \code{NA})}
}
\value{
whether multiple cores are in use (\code{TRUE} or \code{FALSE})
}
\description{
Register a multiprocessing plan with \code{ncores}
}
\examples{
\dontrun{
setup_parallel(2)
}
}
\author{
Kelly Sovacool, \email{sovacool@umich.edu}
}
