% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{shared_ggprotos}
\alias{shared_ggprotos}
\title{Get plot layers shared by \code{plot_mean_roc} and \code{plot_mean_prc}}
\usage{
shared_ggprotos(ribbon_fill = "#D9D9D9", line_color = "#000000")
}
\arguments{
\item{ribbon_fill}{ribbon fill color (default: "#D9D9D9")}

\item{line_color}{line color (default: "#000000")}
}
\value{
list of ggproto objects to add to a ggplot
}
\description{
Get plot layers shared by \code{plot_mean_roc} and \code{plot_mean_prc}
}
\author{
Kelly Sovacool \email{sovacool@umich.edu}
}
