% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_outcome_features}
\alias{split_outcome_features}
\title{Split dataset into outcome and features}
\usage{
split_outcome_features(dataset, outcome_colname)
}
\arguments{
\item{dataset}{Data frame with an outcome variable and other columns as features.}

\item{outcome_colname}{Column name as a string of the outcome variable
(default \code{NULL}; the first column will be chosen automatically).}
}
\value{
list of length two: outcome, features (as dataframes)
}
\description{
Split dataset into outcome and features
}
\examples{
\dontrun{
split_outcome_features(mikropml::otu_mini_bin, "dx")
}
}
\keyword{internal}
