% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stop_parallel}
\alias{stop_parallel}
\title{Unregsiter a parallel processing cluster}
\usage{
stop_parallel(pcluster)
}
\arguments{
\item{pcluster}{a PSOCK cluster created by `setup_cluster``}
}
\description{
Unregsiter a parallel processing cluster
}
\examples{
para_cluster <- setup_parallel(2)
# insert code that uses foreach here
stop_parallel(para_cluster)
}
\author{
Kelly Sovacool, \email{sovacool@umich.edu}
}
