% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_data1}
\alias{test_data1}
\title{A test data partition from \code{otu_medium}}
\format{
An object of class \code{data.frame} with 12 rows and 6921 columns.
}
\usage{
test_data1
}
\description{
A test data partition from \code{otu_medium}
}
\keyword{datasets}
