% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_data_mini}
\alias{test_data_mini}
\title{A test data partition from \code{otu_mini}}
\format{
An object of class \code{data.frame} with 39 rows and 4 columns.
}
\usage{
test_data_mini
}
\description{
A test data partition from \code{otu_mini}
}
\keyword{datasets}
