% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_data_sm}
\alias{test_data_sm}
\title{A test data partition from \code{otu_small}}
\format{
An object of class \code{data.frame} with 12 rows and 61 columns.
}
\usage{
test_data_sm
}
\description{
A test data partition from \code{otu_small}
}
\keyword{datasets}
