% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_hyperparams}
\alias{test_hyperparams}
\title{Test model hyperparameters with less options for faster build}
\format{
a data frame with 23 rows and 3 columns
\describe{
\item{param}{hyperparameter}
\item{value}{value of the hyperparameter}
\item{method}{ML method that the hyperparameter applies to}
}
}
\usage{
test_hyperparams
}
\description{
Test model hyperparameters with less options for faster build
}
\keyword{datasets}
