% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{tidy_perf_data}
\alias{tidy_perf_data}
\title{Tidy the performance dataframe}
\usage{
tidy_perf_data(performance_df)
}
\arguments{
\item{performance_df}{dataframe of performance results from multiple calls to \code{run_ml()}}
}
\value{
Tidy dataframe with model performance metrics
}
\description{
Used by \code{plot_performance()}.
}
\author{
Begüm Topçuoglu, \email{topcuoglu.begum@gmail.com}

Kelly Sovacool, \email{sovacool@umich.edu}
}
