% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{train_data_mini}
\alias{train_data_mini}
\title{A training data partition from \code{otu_mini}}
\format{
An object of class \code{data.frame} with 161 rows and 4 columns.
}
\usage{
train_data_mini
}
\description{
A training data partition from \code{otu_mini}
}
\keyword{datasets}
