% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{train_data_sm}
\alias{train_data_sm}
\title{Small training dataset}
\format{
A data frame with 30 rows and 61 variables.
The \code{dx} column is the diagnosis: healthy or cancerous (colorectal).
All other columns are OTU relative abundances.
}
\usage{
train_data_sm
}
\description{
A training data partition from \code{otu_small}
}
\keyword{datasets}
