% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trained_model_mini}
\alias{trained_model_mini}
\title{A trained model from L2 logistic regression on \code{train_data_mini}}
\format{
An object of class \code{train} (inherits from \code{train.formula}) of length 23.
}
\usage{
trained_model_mini
}
\description{
A trained model from L2 logistic regression on \code{train_data_mini}
}
\keyword{datasets}
