% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning_grid.R
\name{tuning_grid}
\alias{tuning_grid}
\title{Title}
\usage{
tuning_grid(train_data, model, outcome, hyperparameters)
}
\arguments{
\item{train_data}{TODO}

\item{model}{TODO}

\item{outcome}{TODO}

\item{hyperparameters}{TODO}
}
\value{

}
\description{
Title
}
\author{
Begüm Topçuoğlu, \email{topcuoglu.begum@gmail.com}
}
