% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_aucs.R
\name{calc_aucs}
\alias{calc_aucs}
\title{Title}
\usage{
calc_aucs(trained_model, test_data, outcome, fewer_samples)
}
\arguments{
\item{trained_model}{TODO}

\item{test_data}{TODO}

\item{outcome}{TODO}

\item{fewer_samples}{TODO}
}
\value{

}
\description{
Title
}
\author{
Zena Lapp, \email{zenalapp@umich.edu}
}
