% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_aucs.R
\name{calc_aucs}
\alias{calc_aucs}
\title{Calculate AUROC and AUPRC}
\usage{
calc_aucs(trained_model, test_data, outcome, outcome_samples)
}
\arguments{
\item{trained_model}{TODO}

\item{test_data}{TODO}

\item{outcome}{TODO}

\item{outcome_samples}{TODO}
}
\value{

}
\description{
Calculate AUROC and AUPRC
}
\author{
Zena Lapp, \email{zenalapp@umich.edu}
}
