% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_aucs.R
\name{calc_aucs}
\alias{calc_aucs}
\title{Calculate AUROC and AUPRC}
\usage{
calc_aucs(trained_model, test_data, outcome_colname, outcome_value)
}
\arguments{
\item{trained_model}{trained model from caret}

\item{test_data}{dataframe of testing data}

\item{outcome_colname}{Column name of the outcome variable}

\item{outcome_value}{Outcome value of interest}
}
\value{
List of auroc and auprc values
}
\description{
Calculate AUROC and AUPRC
}
\examples{
calc_aucs(trained_model_sm1, test_data_sm, "dx", "cancer")
}
\author{
Zena Lapp, \email{zenalapp@umich.edu}

Kelly Sovacool, \email{sovacool@umich.edu}
}
