% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_aucs.R
\name{calc_auprc}
\alias{calc_auprc}
\title{Calculate AUPRC}
\usage{
calc_auprc(pred, bin_outcome)
}
\arguments{
\item{pred}{TODO}

\item{bin_outcome}{TODO}
}
\value{

}
\description{
Calculate AUPRC
}
\author{
Zena Lapp, \email{zenalapp@umich.edu}
}
