% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_aucs.R
\name{calc_auroc}
\alias{calc_auroc}
\title{Calculate AUROC}
\usage{
calc_auroc(pred, bin_outcomes)
}
\arguments{
\item{pred}{Predictions of the trained model on the test data}

\item{bin_outcomes}{Binary outcome vector}
}
\value{
Area under the receiver-operator characteristic curve
}
\description{
Calculate AUROC
}
\author{
Zena Lapp, \email{zenalapp@umich.edu}
}
