% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_metrics.R
\name{calc_perf_metrics}
\alias{calc_perf_metrics}
\title{Get performance metrics for test data}
\usage{
calc_perf_metrics(
  test_data,
  trained_model,
  outcome_colname,
  perf_metric_function
)
}
\arguments{
\item{test_data}{test data}

\item{trained_model}{trained model}

\item{outcome_colname}{column name as a string of the outcome variable}

\item{perf_metric_function}{function to calculate the performance metric to be used for cross-validation and test performance. Some functions are provided by caret (see \link[caret]{defaultSummary}). Defaults: classification (binary and multi-class) = \code{multiClassSummary}, regression = \code{defaultSummary}}
}
\value{
performance metrics
}
\description{
Get performance metrics for test data
}
\examples{
results <- run_ml(mikropml::otu_mini,'regLogistic',kfold = 2)
calc_perf_metrics(results$test_data, results$trained_model, 'dx',multiClassSummary)
}
