% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_groups}
\alias{check_groups}
\title{Check grouping vector}
\usage{
check_groups(dataset, groups, kfold)
}
\arguments{
\item{dataset}{Dataframe with an outcome variable and other columns as features.}

\item{groups}{Vector of groups to keep together when splitting the data into train and test sets, and for cross-validation; length matches the number of rows in the dataset (default: no groups).}

\item{kfold}{Fold number for k-fold cross-validation (default: \code{5}).}
}
\description{
Check grouping vector
}
\examples{
check_groups(mikropml::otu_mini,
  sample(LETTERS, nrow(mikropml::otu_mini), replace = TRUE),
  kfold = 2
)
}
