% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{collapse_correlated_features}
\alias{collapse_correlated_features}
\title{Remove correlated features}
\usage{
collapse_correlated_features(features)
}
\arguments{
\item{features}{features for ML}
}
\value{
features where perfectly correlated ones are collapsed
}
\description{
Remove correlated features
}
\examples{
collapse_correlated_features(mikRopML::otu_small[, 2:ncol(otu_small)])
}
\author{
Zena Lapp, \email{zenalapp@umich.edu}
}
