% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{combine_hp_performance}
\alias{combine_hp_performance}
\title{Combine hyperparameter performance metrics for several train/test splits}
\usage{
combine_hp_performance(trained_model_lst)
}
\arguments{
\item{trained_model_lst}{list of trained models}
}
\value{
list: dat - dataframe of performance metric for each group of hyperparameters, params - hyperparameters tuned, metric - performance metric used
}
\description{
Combine hyperparameter performance metrics for several train/test splits
}
\examples{
combine_hp_performance(list(otu_mini_results1$trained_model,otu_mini_results1$trained_model))
}
