% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{combine_hp_performance}
\alias{combine_hp_performance}
\title{Combine hyperparameter performance metrics for multiple train/test splits}
\usage{
combine_hp_performance(trained_model_lst)
}
\arguments{
\item{trained_model_lst}{list of trained models}
}
\value{
list: dat - dataframe of performance metric for each group of hyperparameters, params - hyperparameters tuned, metric - performance metric used
}
\description{
Combine hyperparameter performance metrics for multiple train/test splits
}
\examples{
\dontrun{
results <- lapply(seq(100, 102), function(seed) {
  run_ml(otu_small, "glmnet", seed = seed)
  })
models <- lapply(results, function(x) x$trained_model)
combine_hp_performance(models)
}
}
\author{
Zena Lapp, \email{zenalapp@umich.edu}
}
