% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_correlation_matrix.R
\name{compute_correlation_matrix}
\alias{compute_correlation_matrix}
\title{Title}
\usage{
compute_correlation_matrix(
  input_file,
  outcome,
  level,
  cor_value = 1,
  p_value = 0.01
)
}
\arguments{
\item{input_file}{TODO}

\item{outcome}{TODO}

\item{level}{TODO}

\item{cor_value}{TODO}

\item{p_value}{TODO}
}
\value{

}
\description{
Title
}
\author{
Begüm Topçuoğlu, \email{topcuoglu.begum@gmail.com}
}
