% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{default_hyperparams}
\alias{default_hyperparams}
\title{Default model hyperparameters}
\format{
a data frame with 51 rows and 3 columns
\describe{
\item{param}{hyperparameter}
\item{val}{value of the hyperparameter}
\item{model}{model that the hyperparameter applies to}
}
}
\usage{
default_hyperparams
}
\description{
Default model hyperparameters
}
\keyword{datasets}
