% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_cv.R
\name{define_cv}
\alias{define_cv}
\title{Define Cross-Validation Scheme and Training Parameters}
\usage{
define_cv(train_data, outcome_colname)
}
\arguments{
\item{train_data}{Dataframe for training model}

\item{outcome_colname}{Column name of the outcome variable}
}
\value{
Caret object for trainControl that controls cross-validation
}
\description{
Define Cross-Validation Scheme and Training Parameters
}
\examples{
define_cv(train_data_sm, "dx")
}
\author{
Begüm Topçuoğlu, \email{topcuoglu.begum@gmail.com}
}
