% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_val.R
\name{define_cv}
\alias{define_cv}
\title{Define Cross-Validation Scheme and Training Parameters}
\usage{
define_cv(train_data, outcome_colname, kfold = 5, seed = NA)
}
\arguments{
\item{train_data}{training data: dataframe of outcome and features}

\item{outcome_colname}{column name as a string of the outcome variable}

\item{kfold}{fold number for k-fold cross-validation (default: 5)}

\item{seed}{random seed (default: NA)}
}
\value{
Caret object for trainControl that controls cross-validation
}
\description{
Define Cross-Validation Scheme and Training Parameters
}
\examples{
define_cv(train_data_sm, "dx", kfold = 5, seed = 2019)
}
\author{
Begüm Topçuoğlu, \email{topcuoglu.begum@gmail.com}
}
