% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_val.R
\name{define_cv}
\alias{define_cv}
\title{Define Cross-Validation Scheme and Training Parameters}
\usage{
define_cv(
  train_data,
  outcome_colname,
  hyperparams_list,
  perf_metric_function,
  class_probs,
  kfold = 5,
  cv_times = 100,
  group = NULL,
  seed = NA
)
}
\arguments{
\item{train_data}{Dataframe for training model}

\item{outcome_colname}{column name as a string of the outcome variable}

\item{hyperparams_list}{named list of lists of hyperparameters}

\item{perf_metric_function}{function to calculate the performance metric to be used for cross-validation and test performance. Some functions are provided by caret (see \link[caret]{defaultSummary}). Defaults: classification (binary and multi-class) = \code{multiClassSummary}, regression = \code{defaultSummary}}

\item{class_probs}{Whether the \verb{\link[caret]\{trainControl\}} \code{classProbs} argument should be TRUE or FALSE (TRUE for classification, FALSE for regression)}

\item{kfold}{fold number for k-fold cross-validation (default: 5)}

\item{cv_times}{number of partitions to create}

\item{group}{vector of groups to keep together when splitting the data into train and test sets, and for cross-validation; length matches the number of rows in the dataset (default: no groups)}

\item{seed}{random seed (default: NA)}
}
\value{
Caret object for trainControl that controls cross-validation
}
\description{
See \link[caret]{trainControl} for info on how the seed is being set.
}
\examples{
define_cv(train_data_sm,
  outcome_colname = "dx",
  hyperparams_list = get_hyperparams_list(otu_small, "regLogistic"),
  perf_metric_function = caret::twoClassSummary,
  class_probs = TRUE,
  kfold = 5,
  seed = 2019
)
}
\author{
Begüm Topçuoğlu, \email{topcuoglu.begum@gmail.com}
}
