% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_importance.R
\name{find_permuted_auc}
\alias{find_permuted_auc}
\title{Title}
\usage{
find_permuted_auc(model, test_data, outcome, feat, fewer_samples)
}
\arguments{
\item{model}{TODO}

\item{test_data}{TODO}

\item{outcome}{TODO}

\item{feat}{TODO}

\item{fewer_samples}{TODO}
}
\value{

}
\description{
Title
}
\author{
Begüm Topçuoğlu, \email{topcuoglu.begum@gmail.com}
}
