% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_importance.R
\name{find_permuted_auc}
\alias{find_permuted_auc}
\title{Get permuted AUROC difference for a single feature (or group of features)}
\usage{
find_permuted_auc(method, test_data, outcome_colname, feat, outcome_value)
}
\arguments{
\item{method}{ML method. Options: `c("regLogistic", "rf", "rpart2", "svmRadial", "xgbTree")``}

\item{test_data}{held out test data: dataframe of outcome and features}

\item{outcome_colname}{Column name as a string of the outcome variable (default \code{NULL}; will be chosen automatically).}

\item{feat}{feature or group of correlated features to permute}

\item{outcome_value}{Outcome value of interest as a string (default \code{NULL}; will be chosen automatically).}
}
\value{
vector of mean permuted auc and mean difference between test and permuted auc
}
\description{
Get permuted AUROC difference for a single feature (or group of features)
}
\author{
Begüm Topçuoğlu, \email{topcuoglu.begum@gmail.com}

Zena Lapp, \email{zenalapp@umich.edu}
}
