% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_corr_feats.R
\name{flatten_corr_mat}
\alias{flatten_corr_mat}
\title{Flatten correlation matrix to pairs}
\usage{
flatten_corr_mat(cormat)
}
\arguments{
\item{cormat}{correlation matrix computed with stats::cor}
}
\value{
flattened correlation matrix (pairs of features their correlation)
}
\description{
Flatten correlation matrix to pairs
}
\examples{
flatten_corr_mat(data.frame(a=runif(4,-1,1), b=runif(4,-1,1), c=runif(4,-1,1)))
}
\author{
Zena Lapp, \email{zenalapp@umich.edu}
}
