% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_corr_feats.R
\name{flatten_corr_mat}
\alias{flatten_corr_mat}
\title{Flatten correlation matrix to pairs}
\usage{
flatten_corr_mat(cormat)
}
\arguments{
\item{cormat}{correlation matrix computed with stats::cor}
}
\value{
flattened correlation matrix (pairs of features their correlation)
}
\description{
Flatten correlation matrix to pairs
}
\examples{
set.seed(0)
mat <- matrix(runif(100), nrow = 20)
rownames(mat) <- 1:nrow(mat)
colnames(mat) <- 1:ncol(mat)
corr_mat <- stats::cor(mat, method = "spearman")
flatten_corr_mat(corr_mat)
}
\author{
Zena Lapp, \email{zenalapp@umich.edu}
}
