% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_aucs.R
\name{get_bin_outcome}
\alias{get_bin_outcome}
\title{Title}
\usage{
get_bin_outcome(outcome_vec, fewer_samples)
}
\arguments{
\item{fewer_samples}{}
}
\value{

}
\description{
Title
}
\author{
Zena Lapp, \email{zenalapp@umich.edu}
}
