% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{get_caret_processed_df}
\alias{get_caret_processed_df}
\title{Get preprocessed dataframe for continuous variables}
\usage{
get_caret_processed_df(features, method)
}
\arguments{
\item{features}{Dataframe of features for machine learning}

\item{method}{Methods to preprocess the data, described in
\code{\link[caret:preProcess]{caret::preProcess()}} (default: \code{c("center","scale")}, use \code{NULL} for
no normalization).}
}
\value{
Named list:
\itemize{
\item \code{processed}: Dataframe of processed features.
\item \code{removed}: Names of any features removed during preprocessing.
}
}
\description{
Get preprocessed dataframe for continuous variables
}
\examples{
get_caret_processed_df(mikropml::otu_small[, 2:ncol(otu_small)], c("center", "scale"))
}
\author{
Zena Lapp, \email{zenalapp@umich.edu}
}
