% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_corr_feats.R
\name{get_corr_feats}
\alias{get_corr_feats}
\title{Identify correlated features}
\usage{
get_corr_feats(features, cor_value = 1)
}
\arguments{
\item{features}{features used for machine learning}

\item{cor_value}{return correlations above or equal to cor_value (default: 1)}
}
\value{
correlated features
}
\description{
Identify correlated features
}
\examples{
set.seed(0)
mat <- matrix(runif(100), nrow = 20)
rownames(mat) <- 1:nrow(mat)
colnames(mat) <- 1:ncol(mat)
get_corr_feats(mat, 0.4)
}
\author{
Begüm Topçuoğlu, \email{topcuoglu.begum@gmail.com}

Zena Lapp, \email{zenalapp@umich.edu}
}
