% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_feats.R
\name{get_corr_feats}
\alias{get_corr_feats}
\title{Identify correlated features}
\usage{
get_corr_feats(features, corr_thresh = 1, group_neg_corr = TRUE)
}
\arguments{
\item{features}{Features used for machine learning.}

\item{corr_thresh}{For feature importance, group correlations
above or equal to \code{corr_thresh} (range \code{0} to \code{1}; default: \code{1}).}

\item{group_neg_corr}{Whether to group negatively correlated features
together (e.g. c(0,1) and c(1,0)).}
}
\value{
Dataframe of correlated features where the columns are feature1,
feature2, and the correlation between those two features
(anything exceeding corr_thresh).
}
\description{
Identify correlated features
}
\examples{
set.seed(0)
mat <- matrix(runif(100), nrow = 20)
rownames(mat) <- 1:nrow(mat)
colnames(mat) <- 1:ncol(mat)
get_corr_feats(mat, 0.4)
}
\author{
Begüm Topçuoğlu, \email{topcuoglu.begum@gmail.com}

Zena Lapp, \email{zenalapp@umich.edu}
}
