% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_importance.R
\name{get_feature_importance}
\alias{get_feature_importance}
\title{Get feature importance using permutation method}
\usage{
get_feature_importance(
  model,
  train_data,
  test_data,
  outcome_colname,
  outcome_value,
  corr_thresh = 1
)
}
\arguments{
\item{model}{caret model}

\item{train_data}{training data: dataframe of outcome and features}

\item{test_data}{held out test data: dataframe of outcome and features}

\item{outcome_colname}{column name of the outcome}

\item{outcome_value}{outcome value of interest}

\item{corr_thresh}{group correlations above or equal to corr_thresh (default: 1)}
}
\value{
aucs when each feature is permuted, and differences between test auc and permuted auc
}
\description{
Get feature importance using permutation method
}
\author{
Begüm Topçuoğlu, \email{topcuoglu.begum@gmail.com}

Zena Lapp, \email{zenalapp@umich.edu}
}
