% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_importance.R
\name{get_feature_importance}
\alias{get_feature_importance}
\title{Get feature importance using permutation method}
\usage{
get_feature_importance(
  trained_model,
  train_data,
  test_data,
  outcome_colname,
  perf_metric_function,
  perf_metric_name,
  method,
  seed = NA,
  corr_thresh = 1
)
}
\arguments{
\item{trained_model}{trained model from caret}

\item{train_data}{training data: dataframe of outcome and features}

\item{test_data}{held out test data: dataframe of outcome and features}

\item{outcome_colname}{Column name as a string of the outcome variable (default \code{NULL}; will be chosen automatically).}

<<<<<<< HEAD
\item{perf_metric_function}{function to calculate the performance metric to be used for cross-validation and test performance. Some functions are provided by caret (see \link[caret]{defaultSummary}). Defaults: classification (binary and multi-class) = \code{multiClassSummary}, regression = \code{defaultSummary}}

\item{perf_metric_name}{the column name from the output of the function provided to perf_metric_function that is to be used as the performance metric. Defaults: binary classification = \code{"AUC"} (AUROC), multi-class classification = \code{"logLoss"}, regression = \code{"RMSE"}.}
=======
\item{outcome_value}{Outcome value of interest as a string (default \code{NULL}; will be chosen automatically).}
>>>>>>> master

\item{method}{ML method. Options: `c("regLogistic", "rf", "rpart2", "svmRadial", "xgbTree")``}

\item{seed}{Random seed (default: \code{NA}). Your results will be reproducible if you set a seed.}

\item{corr_thresh}{For feature importance, group correlations above or equal to corr_thresh (default: \code{1}).}
}
\value{
dataframe with aucs when each feature is permuted, and differences between test auc and permuted auc
}
\description{
Get feature importance using permutation method
}
\author{
Begüm Topçuoğlu, \email{topcuoglu.begum@gmail.com}

Zena Lapp, \email{zenalapp@umich.edu}
}
