% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_importance.R
\name{get_feature_importance}
\alias{get_feature_importance}
\title{Get feature importance using permutation method}
\usage{
get_feature_importance(
  method,
  train_data,
  test_data,
  outcome_colname,
  outcome_value,
  corr_thresh = 1,
  rseed = NA
)
}
\arguments{
\item{method}{ML method ("regLogistic", "rf", "rpart2", "svmRadial", "xgbTree")}

\item{train_data}{training data: dataframe of outcome and features}

\item{test_data}{held out test data: dataframe of outcome and features}

\item{outcome_colname}{column name as a string of the outcome variable}

\item{outcome_value}{outcome value of interest as a string}

\item{corr_thresh}{for feature importance, group correlations above or equal to corr_thresh (default: 1)}

\item{rseed}{random seed used in \code{run_ml}}
}
\value{
aucs when each feature is permuted, and differences between test auc and permuted auc
}
\description{
Get feature importance using permutation method
}
\author{
Begüm Topçuoğlu, \email{topcuoglu.begum@gmail.com}

Zena Lapp, \email{zenalapp@umich.edu}
}
