% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_feature_importance.R
\name{get_feature_importance}
\alias{get_feature_importance}
\title{Get feature importance using permutation method}
\usage{
get_feature_importance(
  dataset,
  model,
  test_data,
  outcome_colname,
  outcome_value
)
}
\arguments{
\item{dataset}{TODO}

\item{model}{TODO}

\item{test_data}{TODO}

\item{outcome_colname}{TODO}

\item{outcome_value}{TODO}
}
\value{
TODO
}
\description{
Get feature importance using permutation method
}
\author{
Begüm Topçuoğlu, \email{topcuoglu.begum@gmail.com}

Zena Lapp, \email{zenalapp@umich.edu}
}
