% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{get_hp_performance}
\alias{get_hp_performance}
\title{Get hyperparameter peformance metrics}
\usage{
get_hp_performance(trained_model)
}
\arguments{
\item{trained_model}{trained model (e.g. from \code{run_ml()})}
}
\value{
list: dat - dataframe of performance metric for each group of hyperparameters, params - hyperparameters tuned, metric - performance metric used
}
\description{
Get hyperparameter peformance metrics
}
\examples{
get_hp_performance(otu_mini_bin_results_glmnet$trained_model)
}
\author{
Zena Lapp, \email{zenalapp@umich.edu}

Kelly Sovacool \email{sovacool@umich.edu}
}
