% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperparameters.R
\name{get_hyperparams_list}
\alias{get_hyperparams_list}
\title{Set hyperparameters based on ML method and dataset characteristics}
\usage{
get_hyperparams_list(dataset, method)
}
\arguments{
\item{dataset}{dataframe with an outcome variable and other columns as features}

\item{method}{ML method ("regLogistic", "rf", "rpart2", "svmRadial", "xgbTree")}
}
\value{
named list of hyperparameters
}
\description{
Set hyperparameters based on ML method and dataset characteristics
}
\examples{
get_hyperparams_list(otu_mini, "rf")
get_hyperparams_list(otu_medium, "rf")
get_hyperparams_list(otu_mini, "rpart2")
get_hyperparams_list(otu_medium, "rpart2")
}
\author{
Kelly Sovacool, \email{sovacool@umich.edu}
}
