% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{get_outcome_type}
\alias{get_outcome_type}
\title{Get outcome type.}
\usage{
get_outcome_type(outcomes_vec)
}
\arguments{
\item{outcomes_vec}{Vector of outcomes.}
}
\value{
Outcome type (continuous, binary, or multiclass).
}
\description{
If the outcome is numeric, the type is continuous.
Otherwise, the outcome type is binary if there are only two outcomes or
multiclass if there are more than two outcomes.
}
\examples{
get_outcome_type(c(1, 2, 1))
get_outcome_type(c("a", "b", "b"))
get_outcome_type(c("a", "b", "c"))
}
\author{
Zena Lapp, \email{zenalapp@umich.edu}
}
