% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_performance_tbl}
\alias{get_performance_tbl}
\title{Get model performance metrics as a one-row tibble}
\usage{
get_performance_tbl(
  trained_model,
  test_data,
  outcome_colname,
  outcome_value,
  seed = NA
)
}
\arguments{
\item{trained_model}{trained model from caret}

\item{test_data}{held out test data: dataframe of outcome and features}

\item{outcome_colname}{Column name as a string of the outcome variable (default \code{NULL}; will be chosen automatically).}

\item{outcome_value}{Outcome value of interest as a string (default \code{NULL}; will be chosen automatically).}

\item{seed}{Random seed (default: \code{NA}). Your results will be reproducible if you set a seed.}
}
\value{
a one-row tibble with columns \code{cv_auroc}, \code{test_auroc}, \code{test_auprc}, \code{method}, and \code{seed}
}
\description{
Get model performance metrics as a one-row tibble
}
\author{
Kelly Sovacool, \email{sovacool@umich.edu}
}
