% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_performance_tbl}
\alias{get_performance_tbl}
\title{Get model performance metrics as a one-row tibble}
\usage{
get_performance_tbl(
  trained_model,
  test_data,
  outcome_colname,
  perf_metric_function,
  perf_metric_name,
  seed = NA
)
}
\arguments{
\item{trained_model}{trained model}

\item{test_data}{test data}

\item{outcome_colname}{Column name as a string of the outcome variable (default \code{NULL}; will be chosen automatically).}

<<<<<<< HEAD
\item{perf_metric_function}{function to calculate the performance metric to be used for cross-validation and test performance. Some functions are provided by caret (see \link[caret]{defaultSummary}). Defaults: classification (binary and multi-class) = \code{multiClassSummary}, regression = \code{defaultSummary}}

\item{perf_metric_name}{the column name from the output of the function provided to perf_metric_function that is to be used as the performance metric. Defaults: binary classification = \code{"AUC"} (AUROC), multi-class classification = \code{"logLoss"}, regression = \code{"RMSE"}.}
=======
\item{outcome_value}{Outcome value of interest as a string (default \code{NULL}; will be chosen automatically).}
>>>>>>> master

\item{seed}{Random seed (default: \code{NA}). Your results will be reproducible if you set a seed.}
}
\value{
a one-row tibble with columns \code{cv_auroc}, \code{test_auroc}, \code{test_auprc}, \code{method}, and \code{seed}
}
\description{
Get model performance metrics as a one-row tibble
}
\author{
Kelly Sovacool, \email{sovacool@umich.edu}
}
