% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_performance_tbl}
\alias{get_performance_tbl}
\title{Get model performance metrics as a one-row tibble}
\usage{
get_performance_tbl(
  trained_model_caret,
  test_data,
  outcome_colname,
  outcome_value,
  rseed = NA
)
}
\arguments{
\item{rseed}{random seed used in `run_ml``}
}
\value{
a one-row tibble with columns \code{cv_auroc}, \code{test_auroc}, and \code{test_auprc}
}
\description{
Get model performance metrics as a one-row tibble
}
\author{
Kelly Sovacool, \email{sovacool@umich.edu}
}
