% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_aucs.R
\name{get_pred}
\alias{get_pred}
\title{Title}
\usage{
get_pred(trained_model, test_data, fewer_samples)
}
\arguments{
\item{fewer_samples}{}
}
\value{

}
\description{
Title
}
\author{
Zena Lapp, \email{zenalapp@umich.edu}
}
