% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_val.R
\name{get_seeds_trainControl}
\alias{get_seeds_trainControl}
\title{Get seeds for caret::trainControl}
\usage{
get_seeds_trainControl(kfold, cv_times, ncol_train)
}
\arguments{
\item{kfold}{}

\item{cv_times}{Number of partitions to create}
}
\value{
seeds for \code{caret::trainControl}
}
\description{
Get seeds for caret::trainControl
}
