% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_val.R
\name{get_seeds_trainControl}
\alias{get_seeds_trainControl}
\title{Get seeds for caret::trainControl}
\usage{
get_seeds_trainControl(hyperparams_list, kfold, cv_times, ncol_train)
}
\arguments{
\item{hyperparams_list}{named list of lists of hyperparameters}

\item{kfold}{fold number for k-fold cross-validation (default: 5)}

\item{cv_times}{Number of partitions to create}

\item{ncol_train}{number of columns in training data}
}
\value{
seeds for \code{caret::trainControl}
}
\description{
Adapted from \href{https://stackoverflow.com/a/32598959}{this Stack Overflow post}
and the \link[caret]{trainControl} documentation
}
\examples{
get_seeds_trainControl(get_hyperparams_list(otu_small, 'regLogistic'), 5, 100, 60)
}
