% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuning_grid.R
\name{get_tuning_grid}
\alias{get_tuning_grid}
\title{Generate the Tuning Grid for Tuning Hyperparameters}
\usage{
get_tuning_grid(hyperparams_df, method)
}
\arguments{
\item{hyperparams_df}{Dataframe with columns \code{param} and \code{value}}

\item{method}{ML method ("regLogistic", "svmRadial", "rpart2", "rf", "xgbTree")}
}
\value{
The tuning grid
}
\description{
Generate the Tuning Grid for Tuning Hyperparameters
}
\examples{
mikRopML:::check_hyperparams_df(default_hyperparams, "regLogistic") \%>\%
  get_tuning_grid()
}
\author{
Begüm Topçuoğlu, \email{topcuoglu.begum@gmail.com}

Kelly Sovacool, \email{sovacool@umich.edu}
}
