% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_feature_importance.R
\name{group_correlated_features}
\alias{group_correlated_features}
\title{Group correlated features}
\usage{
group_correlated_features(corr, test_data)
}
\arguments{
\item{corr}{output of get_corr_feats (pairs of correlated features)}

\item{test_data}{test data from machine learning}
}
\value{
vector of correlated features where each element is the group of correlated features separated by pipes (|)
}
\description{
Group correlated features
}
\author{
Begüm Topçuoğlu, \email{topcuoglu.begum@gmail.com}

Zena Lapp, \email{zenalapp@umich.edu}
}
