% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mikropml.R
\docType{package}
\name{mikropml}
\alias{mikropml}
\title{mikropml: User-Friendly R Package for Robust Machine Learning Pipelines}
\description{
\code{mikropml} implements robust machine learning pipelines using regression,
support vector machines, decision trees, random forest, or gradient-boosted trees.
The main functions are \code{preprocess_data()} to process your data prior to
running machine learning, and \code{run_ml()} to run machine learning.
}
\section{Authors}{

\itemize{
\item Begüm D. Topçuoğlu (\href{http://orcid.org/0000-0003-3140-537X}{ORCID})
\item Zena Lapp (\href{http://orcid.org/0000-0003-4674-2176}{ORCID})
\item Kelly L. Sovacool (\href{http://orcid.org/0000-0003-3283-829X}{ORCID})
\item Evan Snitkin (\href{http://orcid.org/0000-0001-8409-278X}{ORCID})
\item Jenna Wiens (\href{http://orcid.org/0000-0002-1057-7722}{ORCID})
\item Patrick D. Schloss (\href{http://orcid.org/0000-0002-6935-4275}{ORCID})
}
}

\section{See also}{

\itemize{
\item \href{http://www.schlosslab.org/mikropml/articles/}{All vignettes}.
\item \href{http://www.schlosslab.org/mikropml/articles/introduction.html}{Intro vignette}.
\item \href{http://www.schlosslab.org/mikropml/articles/parallel.html}{Parallel vignette} to make it run faster.
\item \href{https://github.com/SchlossLab/mikropml-snakemake-workflow}{Snakemake workflow} for high-performance computing clusters.
\item \href{http://www.schlosslab.org/mikropml/articles/paper.html}{The mikropml paper}.
}
}

