% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mikropml.R
\docType{package}
\name{mikropml}
\alias{mikropml}
\title{mikropml: User-Friendly R Package for Robust Machine Learning Pipelines}
\description{
TODO longer description (paragraph) goes here
}
\section{TODO}{
 brief overview of functions

The heart of \code{mikropml} is \code{run_ml()}.
TODO: brief overview of \code{run_ml()} and \code{preprocess_data()}, then link to vignettes.
}

