% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mikropml.R
\docType{package}
\name{mikropml}
\alias{mikropml}
\title{mikropml: User-Friendly R Package for Robust Machine Learning Pipelines}
\description{
\code{mikropml} implements supervised machine learning pipelines using regression,
support vector machines, decision trees, random forest, or gradient-boosted trees.
The main functions are \code{preprocess_data()} to process your data prior to
running machine learning, and \code{run_ml()} to run machine learning.
}
\section{Authors}{

\itemize{
\item Begüm D. Topçuoğlu (\href{https://orcid.org/0000-0003-3140-537X}{ORCID})
\item Zena Lapp (\href{https://orcid.org/0000-0003-4674-2176}{ORCID})
\item Kelly L. Sovacool (\href{https://orcid.org/0000-0003-3283-829X}{ORCID})
\item Evan Snitkin (\href{https://orcid.org/0000-0001-8409-278X}{ORCID})
\item Jenna Wiens (\href{https://orcid.org/0000-0002-1057-7722}{ORCID})
\item Patrick D. Schloss (\href{https://orcid.org/0000-0002-6935-4275}{ORCID})
}
}

\section{See vignettes}{

\itemize{
\item \href{http://www.schlosslab.org/mikropml/articles/introduction.html}{Introduction}
\item \href{http://www.schlosslab.org/mikropml/articles/preprocess.html}{Preprocessing data}
\item \href{http://www.schlosslab.org/mikropml/articles/tuning.html}{Hyperparameter tuning}
\item \href{http://www.schlosslab.org/mikropml/articles/parallel.html}{Parallel processing}
\item \href{http://www.schlosslab.org/mikropml/articles/paper.html}{The mikropml paper}
}
}

