% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_pipeline.R
\name{model_pipeline}
\alias{model_pipeline}
\title{Title}
\usage{
model_pipeline(
  data,
  model,
  split_number,
  outcome = NA,
  hyperparameters = NA,
  level = NA,
  permutation = TRUE
)
}
\arguments{
\item{permutation}{}
}
\value{

}
\description{
Title
}
\author{
Begüm Topçuoğlu, \email{topcuoglu.begum@gmail.com}
}
