% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{otu_mini_bin_results_rf}
\alias{otu_mini_bin_results_rf}
\title{Results from running the pipline with random forest on \code{otu_mini}}
\format{
An object of class \code{list} of length 4.
}
\usage{
otu_mini_bin_results_rf
}
\description{
Results from running the pipline with random forest on \code{otu_mini}
}
\keyword{datasets}
