% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation_importance.R
\name{permutation_importance}
\alias{permutation_importance}
\title{Title}
\usage{
permutation_importance(
  dataset,
  model,
  test_data,
  outcome_colname,
  outcome_value
)
}
\arguments{
\item{dataset}{TODO}

\item{model}{TODO}

\item{test_data}{TODO}

\item{outcome_colname}{TODO}

\item{outcome_value}{TODO}
}
\value{
TODO
}
\description{
Title
}
\author{
Begüm Topçuoğlu, \email{topcuoglu.begum@gmail.com}

Zena Lapp, \email{zenalapp@umich.edu}
}
