% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{preprocess_data}
\alias{preprocess_data}
\title{Preprocess data prior to running machine learning}
\usage{
preprocess_data(
  dataset,
  outcome_colname,
  method = c("center", "scale"),
  remove_nzv = TRUE,
  collapse_corr_feats = TRUE,
  to_numeric = TRUE,
  group_neg_corr = TRUE
)
}
\arguments{
\item{dataset}{dataframe where rows are samples and colums are the outcome variable and features}

\item{outcome_colname}{column name as a string of the outcome variable}

\item{method}{methods to preprocess the data, described in \code{caret::preProcess} (defaut: \code{c("center","scale")}, use \code{NULL} for no normalization)}

\item{remove_nzv}{whether to remove variables with near-zero variance (default: \code{TRUE})}

\item{collapse_corr_feats}{whether to keep only one of perfectly correlated featurs}

\item{to_numeric}{whether to change features to numeric where possible}

\item{group_neg_corr}{whether to group negatively correlated features together (e.g. c(0,1) and c(1,0))}
}
\value{
preprocessed data
}
\description{
Preprocess data prior to running machine learning
}
\examples{
preprocess_data(mikropml::otu_small, "dx")
}
\author{
Zena Lapp, \email{zenalapp@umich.edu}
}
