% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{process_novar_feats}
\alias{process_novar_feats}
\title{Process features with no variation}
\usage{
process_novar_feats(features)
}
\arguments{
\item{features}{dataframe of features for machine learning}
}
\value{
list of two dataframes: features with variability (unprocessed) and without (processed)
}
\description{
Process features with no variation
}
\examples{
process_novar_feats(mikRopML::otu_small[,2:ncol(otu_small)])
}
