% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{rm_corr_feats}
\alias{rm_corr_feats}
\title{Remove correlated features}
\usage{
rm_corr_feats(features, group_neg_corr = TRUE)
}
\arguments{
\item{features}{features for ML}

\item{group_neg_corr}{whether to group negatively correlated features together (e.g. c(0,1) and c(1,0))}
}
\value{
features where perfectly correlated ones are collapsed
}
\description{
Remove correlated features
}
\examples{
rm_corr_feats(mikropml::otu_small[, 2:ncol(otu_small)])
}
\author{
Zena Lapp, \email{zenalapp@umich.edu}
}
