% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ml.R
\name{run_ml}
\alias{run_ml}
\title{Run the machine learning pipeline}
\usage{
run_ml(
  dataset,
  method,
  outcome_colname = NA,
  outcome_value = NA,
  hyperparameters = mikRopML::default_hyperparams,
  find_feature_importance = FALSE,
  kfold = 5,
  training_frac = 0.8,
  corr_thresh = corr_thresh,
  seed = NA,
  ncores = NA
)
}
\arguments{
\item{dataset}{dataframe with an outcome variable and other columns as features}

\item{method}{ML method ("regLogistic", "svmRadial", "rpart2", "rf", "xgbTree")}

\item{outcome_colname}{column name as a string of the outcome variable}

\item{outcome_value}{outcome value of interest as a string}

\item{hyperparameters}{dataframe of hyperparameters (default: default_hyperparams)}

\item{find_feature_importance}{run permutation imporance (default: FALSE)}

\item{kfold}{fold number for k-fold cross-validation (default: 5)}

\item{training_frac}{fraction size of data for training (default: 0.8)}

\item{corr_thresh}{for feature importance, group correlations above or equal to corr_thresh (default: 1)}

\item{seed}{random seed (default: NA)}

\item{ncores}{number of cores for parallel processing (default: NA). \code{parallel} and \code{doParallel} packages are needed for ncores > 1}
}
\value{
named list with results
}
\description{
Run the machine learning pipeline
}
\author{
Begüm Topçuoğlu, \email{topcuoglu.begum@gmail.com}

Zena Lapp, \email{zenalapp@umich.edu}

Kelly Sovacool, \email{sovacool@umich.edu}
}
