% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pipeline.R
\name{run_pipeline}
\alias{run_pipeline}
\title{Run machine learning pipeline}
\usage{
run_pipeline(
  dataset,
  model,
  outcome_colname = NA,
  outcome_value = NA,
  hyperparameters = default_hyperparams,
  permute = FALSE,
  seed = NA
)
}
\arguments{
\item{dataset}{TODO}

\item{model}{TODO}

\item{outcome_colname}{TODO}

\item{outcome_value}{TODO}

\item{hyperparameters}{TODO}

\item{permutation}{TODO}
}
\value{

}
\description{
Run machine learning pipeline
}
\author{
Begüm Topçuoğlu, \email{topcuoglu.begum@gmail.com}
}
