% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pipeline.R
\name{run_pipeline}
\alias{run_pipeline}
\title{Run machine learning pipeline}
\usage{
run_pipeline(
  dataset,
  method,
  outcome_colname = NA,
  outcome_value = NA,
  hyperparameters = mikRopML::default_hyperparams,
  metric = "ROC",
  permute = FALSE,
  seed = NA
)
}
\arguments{
\item{dataset}{TODO}

\item{method}{TODO}

\item{outcome_colname}{TODO}

\item{outcome_value}{TODO}

\item{hyperparameters}{TODO}

\item{metric}{TODO}

\item{permute}{TODO}

\item{seed}{random seed (default: NA)}
}
\value{
named list with results
}
\description{
Run machine learning pipeline
}
\author{
Begüm Topçuoğlu, \email{topcuoglu.begum@gmail.com}
}
