% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_feature_importance.R
\name{select_apply}
\alias{select_apply}
\title{Use future apply if available}
\usage{
select_apply(fun = "apply", ...)
}
\arguments{
\item{fun}{apply function to use (apply, lapply, sapply, etc.)}

\item{...}{all arguments to input to the apply function (in the correct order)}
}
\value{

}
\description{
Use future apply if available
}
\examples{
select_apply(fun='sapply',1:10,function(x) x*10)
}
