% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setup_parallel}
\alias{setup_parallel}
\title{Register a cluster for parallel processing}
\usage{
setup_parallel(ncores, setup_timeout = 0.5)
}
\arguments{
\item{ncores}{number of cores for parallel processing (default: NA). \code{parallel} and \code{doParallel} packages are needed for ncores > 1}

\item{setup_timeout}{Setup timeout in seconds. See \link[parallel]{makePSOCKcluster} for details.}
}
\value{
PSOCK cluster, or NULL if only 1 core provided or required packages aren't available
}
\description{
Register a cluster for parallel processing
}
\examples{
para_cluster <- setup_parallel(2)
# insert code that uses foreach here
stop_parallel(para_cluster)
}
\author{
Kelly Sovacool, \email{sovacool@umich.edu}
}
