% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_outcome_features}
\alias{split_outcome_features}
\title{Split dataset into outcome and features}
\usage{
split_outcome_features(dataset, outcome_colname)
}
\arguments{
\item{dataset}{dataframe with an outcome variable and other columns as features}

\item{outcome_colname}{column name as a string of the outcome variable}
}
\value{
list of length two: outcome, features
}
\description{
Split dataset into outcome and features
}
\examples{
split_outcome_features(mikRopML::otu_mini,'dx')
}
