% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{train_data1}
\alias{train_data1}
\title{Results from running the pipline with L2 logistic regression on \code{otu_mini}
A training data partition from \code{otu_medium}}
\format{
An object of class \code{data.frame} with 48 rows and 6921 columns.
}
\usage{
train_data1
}
\description{
Results from running the pipline with L2 logistic regression on \code{otu_mini}
A training data partition from \code{otu_medium}
}
\keyword{datasets}
