% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{train_data_sm}
\alias{train_data_sm}
\title{A training data partition from \code{otu_small}}
\format{
An object of class \code{data.frame} with 48 rows and 61 columns.
}
\usage{
train_data_sm
}
\description{
A training data partition from \code{otu_small}
}
\keyword{datasets}
