% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_model.R
\name{train_model}
\alias{train_model}
\title{Train model}
\usage{
train_model(
  model_formula,
  train_data,
  method,
  cv,
  perf_metric_name,
  tune_grid,
  ntree
)
}
\arguments{
\item{model_formula}{model formula}

\item{train_data}{train_data}

\item{method}{method}

\item{cv}{cross-validation caret scheme}

\item{perf_metric_name}{performance metric name}

\item{tune_grid}{tuning grid}

\item{ntree}{number of trees for random forest}
}
\value{
trained model
}
\description{
Train model
}
